getLanguage of the QueryObjectModel returns JCR_SQL2 instead of JCR_JQOM
    From the specs:

String Query.getLanguage()

returns the language in which the query is specified.

If the Query was created with an explicitly supplied language string parameter using QueryManager.createQuery then this method returns that string.

If the Query is actually a QueryObjectModel created with QueryObjectModelFactory.createQuery then Query.getLanguage will return the string constant Query.JCR_SQL2.